Compiles, finally. Ran into an issue with recursive mutex's, think that's licked. Fixed endian issues.
Outstanding: Need to seperate linux.x86 from mach.x86, utterly unsure of how to do that.
When compiling test, ld complains about -lrt (not needed probably, or the getclock functionality is in a different library?) and -lif (not even sure what this is, and Google searches on possible terms including libif and iflib return nothing relevent).