Adds another lpc1768 target, using the lpcopen driver library instead of mbed. This new target resembles in naming the lpc1769 target (starts with lpc instead of mbed).
Fixes the hard_fault_handler on the lpc17xx_generic's startup.c to work with gdb backtracing.
Fixes the clock speed on the lpc17xx, because based on how the code runs it comes out at 96 MHz and not 100.
Fixes link errors in the lm4f120xl launchpad.
Adds configurable CAN bitrate feature to the LPC CAN driver.