Move Tivaware link options to etc/tivaware.mk
Moves the link option required when linking to the Tivaware libaries to etc/tivaware.mk.  Previously these were local to the Makefiles.  This provides a common configuration to switch from using the prebuilt libraries to those generated locally.

The freertos_tcp.mk fiel was also updated to provide similar functionality for the freertos_tcp libraries.