more work on the ESP32 port.
removing a number of ESP32 checks in favor of requiring __FreeRTOS__
being defined. By doing this the required file modifications are
slightly smaller but there are still a number of path issues in the
includes and a few pieces that are different on the older FreeRTOS stack
used by the ESP32. The two examples mostly compile now but there are
link errors on SerialBridge which I haven't been able to resolve just
yet. Notes in the readme.md file updated with current state, this will
be updated to have more documentation on the Arduino port as it gets
more mature.