Initial ESP32 port based on the Arduino stack.
The ESP32 port is somewhat functional, there is an issue with the
SerialBridge class not having a proper loop_for_read implementation
(@balazsracz please review).

Two examples have been provided:
ESP32SerialBridge - This does not compile due to HardwareSerial not
matching signature with SerialBridge usage (to be resolved still).
ESP32WiFiClientBridge - This compiles and runs but JMRI doesn't seem to
be able to interact with the node, most likely due to the loop_for_read
implementation being broken.