Split MDNS service name to have protocol as its own constant.
On the ESP32 the mDNS subsystem uses split protocol/service names as
does AvaHi. To ensure the uniform mDNS broadcast by OpenMRN compatible
devices I've split the constants for easier adoption. The only exception
to this is currently with SocketClient (used in modebus_controller
custom-cc3220sf) which will need some further refinement to accept the
protocol as a second parameter and assemble the final service lookup
string internally. I've not covered that update at this time in favor of
updating main.cxx to assemble the string to pass in (I don't have that
hardware to test with)