Deferred stack startup
On the ESP32 (and other Arduino stacks) it is required to defer the
OpenMRN stack startup until the setup() method has been invoked. With
this commit I've split the init(nodeid) method to only initialize the
stack but not start it. It will now be required to call the
OpenMRN.begin() method as part of the setup() method to start the
OpenMRN stack after initializing any required subsystems.

Additionally with this change is the dependency on
SNIP_DYNAMIC_FILENAME. Without this being defined the examples will fail
to compile.