ESP32 hardware CAN driver support
The ESP-IDF hardware CAN driver is now working with an MCP2551
transceiver connected to the ESP32. It is possible to chain multiple
ESP32 together on one CAN bus and have them all show up via connecting
to one of the nodes in JMRI.

Added documentation to various methods for the ESP32 code.