Adding factory reset helper to examples
When pulling up the configuration dialog for the node the first time in
JMRI the name and description fields were showing up with garbage
characters (non-printable) since the fields were not properly
initialized on the node. With the factory reset helper in place this is
now properly displaying in the CDI by default.