Removing the background task and updating examples for review comments.
The background task is an added complexity that is not necessary at this
time. If a consumer of the library wants to move the openmrn.loop() call
to its own task they can do that in their code instead. The ESP32 CAN
RX/TX tasks will remain as-is using the default priority defined in
OpenMRN.h as a starting point for their priority.

Updating the examples to remove references to the background task,
removing reference to DevKit as it is not used on the esp32 and
adjusting the factory reset callback to use the CDI_FACTORY_RESET method
where appropriate.