updates to the powering section and adding a new GPIO section.
The powering the esp32 section needed a lot of clarifications and
rework, it is now more clear as to how/when you can connect the esp32 to
the CAN bus for power or a standalone power supply. The ESP32 is a
relatively power hungry MCU and can not handle the default power
supplied via the CAN bus.

Additionally added a new section about a handful of GPIO pins that have
some restrictions on use.