Fixing ESP32 IO Board example
The GPIO pins were not being initialized or setup in a polling loop, so
in effect the GPIO events could never be processed.

Removed a call to set_gpio() in ArduinoGpio.hxx as this call is not
required with the usage of the pinMode function. This call may be
restored if an ESP32 specific form of the ArduinoStaticGpio base class
is created using lower level APIs.