ESP32 WiFi manager code
This introduces an ESP32 WiFi manager, this allows a node to be either a
hub or a client (or both!). This will simplify the ESP32 nodes
considerably as it will allow all WiFi details to be managed
automatically and expose some configuration parameters via CDI.

The WiFi stack is by default managed by the Esp32WiFiManager code but
can be managed by the caller by passing a null value for ssid, password
and hostname_prefix. When this mode of operation is in use it is
required that the caller has already started the WiFi stack, connected
to the SSID and started the mDNS service. This is not generally required
for most node types.

It is required to have a WiFiConfiguration CDI_GROUP declared and
passed into the constructor for the Esp32WiFiManager class. The
Esp32WiFiManager class will register itself as a configuration
listener and restart the background task if the configuration has
been updated.

The Esp32WiFiManager uses one background task when not operating as a
hub. If the hub mode is enabled a second task will be started internally
to accept connections on the configured TCP/IP socket. The background
task started by the Esp32WiFiManager monitors any established TCP/IP
connections. For inbound connections (when hub mode is active) if they
are no longer connected they will be cleaned up. For outbound
connections one attempt will be made to reconnect, if that attempt fails
the connection will be cleaned up. In addition to monitoring the
connection health, when the "Search mode" is set to one of the auto
modes an mDNS scan will be performed roughly every 30 seconds.

When the "Search mode" is configured with one of the "manual" options
selected, the configuration is checked to ensure that the manual
configuration settings have been configured. If they have not the
"Search mode" will be reset to "auto only".