review comments and experimental GcTcpHub usage
simplified the Esp32WiFiManager constructor args and introduced a second
one which handles the "unmanaged" WiFi/MDNS use-case.

The apply_configuration method now uses an MD5 of the persistent config
to check for updated configuration and when there is a difference (or
initial_load is set) a flag is set for the wifi_manager_task to reload
configuration on its next iteration.

wifi_manager_task is primarily in a delay state and will wake up roughly
every 30sec. When it wakes up it will check if the reload flag is set
and when set it will shutdown the hub (if enabled previously) and
disconnect from its uplink hub. It will then load the config and if the
hub is enabled it will start a GcTcpHub instance and advertise the MDNS
name configured in the CDI. After config check is complete the task will
connect to an uplink hub if it can (auto/manual/last-hub/etc). Only one
uplink connection will be established from the node in an effort to
ensure it doesn't cause the node to create a loop between two
inter-connected nodes.

Esp32WiFiClientAdapter has been updated to support reconnect and
disconnect (stop). However, disconnecting from a remote connection may
result in a null deref inside WiFiClient which I have not narrowed down
fully yet. I'm looking at removing WiFiClient usages to avoid this (and
other) issues which have been observed with it's usage.

At this point the ESP32 "hub" mode of operation should be considered
experimental and not ready for production deployment. But as a leaf node
connecting to another node that is a hub does appear to be safe.