further work on the esp32 as a hub.
shifted to using GcTcpHub with select mode enabled, reimplemented the
OS.hxx/os.h/os.c code to share almost all of the FreeRTOS code. Added
OS.cxx with helper functions used on the ESP32 platform since the
vTaskSetApplicationTaskTag method is not enabled by default on the
ESP32. There is an issue in the client connection to remote hubs where
it does not always detect that the connection has died, this appears to
be a bug in the ESP32 WiFiClient abstraction layer which has limited
usage and is going to be replaced with generic socket calls (via
SocketClient likely) and the wrapper class being removed.