Dead connection checking improvements
Moved dead connection checking to just after config update detection,
this way a dead connection that can not be re-established will be
discarded and a new connection can be re-attempted after rediscovery
via mDNS (if enabled) or manual connection (if enabled). Further
refinements could be made as the node could remain in a disconnected
state for up to 30sec (connection check interval).