ESP32 WiFi Manager
This brings in a WiFi Manager for the ESP32 as well as allows the ESP32
to be a hub. Using this approach it simplifies the ESP32 application
code so that it doesn't need to manage the WiFi/mDNS stacks or the
hub/uplink connections. The application can optionally manage the
WiFi/mDNS stack if it needs to but this is not expected to be a common
usecase.