Ensure newline is sent for LOG(FATAL, ...) before abort()
When LOG(FATAL, ...) is used on the ESP32 the message gets lost since
there is no EOL character before the abort to trigger flush of STDERR.
Adding a newline immediately after the fprintf call forces a flush of
STDERR prior to abort() being called.

Also fixing a couple typos in the comments.