mDNS timeout adjustment.
Lowering mDNS query timeout from 3000ms to 2000ms as JMRI fails
consistently to load the CDI from the ESP32 when it is a HUB and mDNS
query is running. This is consistently failing at 2750ms+. The default
timeout in the Arduino MDNS library is 3000ms and is not configurable.
In the ESP-IDF examples there is a mix of 2000ms and 3000ms for various
query operations so the reduction to 2000ms is not unreasonable. More
work is to be done to see if this can be increased further.