Update examples after further testing
After further testing the overrides that were in place for the older
WiFiClientAdapter based implementation are causing unnecessary delays in
delivery of GC packet data. Added one override (commented) to enable the
usage of select(), there doesn't appear to be a significant performance
difference when select() is enabled vs disabled.