ESP32 WiFi Manager (#249)
- Adds an ESP32 Wifi Manager that uses lowlevel esp32 APIs to connect to wifi instead of the arduino libraries. Dramatically reduces the number of lines of code in the sketch itself.
- Removes the esp32 wifi client adapter.
- Makes hub and uplink configurable via CDI in the wifi manager code. Uses socket client for uplink.
- Automatically monitors the wifi conneciton state and restarts the hub / uplink when wifi gets dropped and reconnected.
- adds support for mdns through the openmrn abstraction layer
- uses the same symbols for specifying wifi ssid and password as the cc32xx
- makes SocketClient compatible with ESP32 tcp/ip system

misc fixes:
* exclude vscode generated artifacts
* Ensure newline is sent for LOG(FATAL, ...) before abort()
* Refactors OS-specific conditionals in HubDevice to use openmrn_features.h comments.
* removing references to Arduino.h where possible



Including Arduino.h is problematic as it polutes the namespaces which
will cause conflicts when including other code. Dropped Arduino.h where
possible and swapped for a more narrow scoped include in ArduinoGpio for
the ESP32 (still wider than I'd like but including esp32-hal-gpio.h
includes esp32-hal.h)

* ESP32 WiFi Manager

This brings in a WiFi Manager for the ESP32 as well as allows the ESP32
to be a hub. Using this approach it simplifies the ESP32 application
code so that it doesn't need to manage the WiFi/mDNS stacks or the
hub/uplink connections. The application can optionally manage the
WiFi/mDNS stack if it needs to but this is not expected to be a common
usecase.

* Add SYSTEM_EVENT_STA_LOST_IP and clarify a couple log messages

* exclude vscode generated artifacts

* Ensure newline is sent for LOG(FATAL, ...) before abort()

When LOG(FATAL, ...) is used on the ESP32 the message gets lost since
there is no EOL character before the abort to trigger flush of STDERR.
Adding a newline immediately after the fprintf call forces a flush of
STDERR prior to abort() being called.

Also fixing a couple typos in the comments.

* Documentation updates and connection status improvements

Added documentation to the ESP32 CDI elements for WiFi.

Split the GOTIP event group bit into SSID connected and GOTIP so that
tracking is a bit more accurate. Adjusted connection/IP timeout from
200s to 180s as part of this.

Added SYSTEM_EVENT_STA_CONNECTED event handling since this is raised
when the esp_wifi_connect() call completes the connection to the SSID.

Added more documentation to wifi_manager_task as part of the cleanup.

* Cleanup of header includes

* mDNS timeout adjustment.

Lowering mDNS query timeout from 3000ms to 2000ms as JMRI fails
consistently to load the CDI from the ESP32 when it is a HUB and mDNS
query is running. This is consistently failing at 2750ms+. The default
timeout in the Arduino MDNS library is 3000ms and is not configurable.
In the ESP-IDF examples there is a mix of 2000ms and 3000ms for various
query operations so the reduction to 2000ms is not unreasonable. More
work is to be done to see if this can be increased further.

* Removing WiFiClientAdapter

Removing the Esp32WiFiClientAdapter code as it is no longer used and is
now redundant with the Esp32WiFiManager code in place.

* Update examples after further testing

After further testing the overrides that were in place for the older
WiFiClientAdapter based implementation are causing unnecessary delays in
delivery of GC packet data. Added one override (commented) to enable the
usage of select(), there doesn't appear to be a significant performance
difference when select() is enabled vs disabled.

* Initial work towards a common WiFiInterface for CC32xx and ESP32.

* Updates for consistency, stringify, const and reference params, etc

Renamed InterfaceMode to WiFiMode

switched from const char * to const std::string & for params where
possible.

modified setup_access_point to default to OPEN with only required param
being the SSID to advertise. I'm not sure this is ideal, perhaps
switching to SSID:OpenMRN, security_key:OpenMRN, security_type:WPA2?

Small clarifications on documentation.

Added WiFiMode to get_ip() and get_mac() to support WiFi interfaces that
support more than one IP or MAC address.

* Namespace updates for esp32, reuse common WIFI_SSID/PASS defines

Wrapping the Esp32WiFiManager code in namespace openmrn_arduino, only
the MDNS and "missing" standard lib functions are left in the global
namespace.

Replacing DEFAULT_WIFI_NAME and DEFAULT_WIFI_PASSWORD with WIFI_SSID and
WIFI_PASS from WiFiDefs.hxx to not have different namings for these
parameters between different implementations.

* Review comments

Adding WiFiErrorCode enum for various result codes for the APIs, adding
blocking param to APIs which could be blocking along with corresponding
timeout value.

* Review comments.

1) Splitting WPS API to support PBC and PIN.
2) Adding BSSID and channel to NetworkListEntry.
3) Adding API for continuous network scan.
4) Rename APIs to be <subject>_<action> with exception of get_mode and
get_state().

* Fix mDNS port value.

Adding missing htons call for the port returned from mDNS as well as
print the port in the logs

* Clarify IDLE as a default state from start().

* remove () in refs

* remove WiFiInterface.hxx so it can be its own PR later

* review feedback.

1) switched examples to use one_two instead of oneTwo for global vars.
2) added __MACH__ for SIGPIPE and EOF reporting.
3) added _FEATURE_ to SIGPIPE and EOF defines.
4) removed using line for WiFiConfiguration CDI element.

* add reserved space in CDI element

* further review comments

1) code reformat.
2) WIFI_TASK_PRIORITY increased from one to two.
3) HUB_STARTUP_DELAY switched from ticks to uS.
4) OVERRIDE -> override.
5) simplify password copy in start_wifi_system.
6) convert SPLIT_MDNS_SERVICE_NAME to split_mdns_service_name method.

* Fixes some minor comments.

* Fix capitalization typo.

* more <> vs "" fixes