Improve Esp32 WiFi SSID disconnect event handler.
Originally the Esp32WiFiManager would only react to the
SYSTEM_EVENT_STA_DISCONNECTED event if we had previously connected. In
some cases this may not be correct, such as SSID not found or one of a
few timeout cases. The updated code now prints why the
SYSTEM_EVENT_STA_DISCONNECTED event was raised as well as attempts to
reconnect if it is not an SSID not found error which will result in a
FATAL log message which triggers a node reboot.