combine AP_NOT_FOUND and ASSOC_EXPIRE to reinit stack.
After testing against a different AP with weaker signal (among other
factors) I consistently failed to connect to the AP on first attempt but
after a restart of the node (via abort) it connected every time. I've
now adjusted the disconnected event to handle this in the same manner as
ASSOC_EXPIRE which forces a reinit of the WiFi stack which allows the
node to connect to the AP without the restart.