Adjust reconnection flow for full reinit
When we have cached data in memory unless we fully deinitialize the WiFi
stack it appears that it retains the cache even though we stop and start
the stack. Adding a call to deinit the stack will force purge the cached
data allowing a clean reinit. The only time reinit is currently
triggered is when we get back a disconnect event for AUTH_EXPIRE or
ASSOC_EXPIRE, all other cases will be handled as simple reconnect
requests.

Added a 250ms delay from when we get disconnected to when we attempt to
reconnect to the SSID and moved AP_NOT_FOUND to default case of
reconnect without reinit.