TCP interface (#262)
Reverts "Revert \"TCP interface\""
Fixes compilation problems.

Original description: 

Implements an interface using the draft Tcp protocol. 
The new interface operates in conjunction with a Hub<string>, where each TCP rendered packet is in exactly one buffer traversing the Hub. A select-aware state flow is added that segments incoming bytes from an fd into packets of the Tcp protocol format with as little memory fragmentation as possible. This needed a new HubDeviceSelect (which is now templated on the read segmenting flow) and the equivalent code of create_gc_port_for_can_hub to handle graceful closing of sockets by the remote side, socket errors or teardown of the interface locally (in unit tests).
This scheme supports having multiple clients connected to a single hub, which is necessary for TCP servers. There is no filtering, i.e., all messages are passed on to every client.

The tcp interface itself converts between the Buffer<string> having the binary (wire) representation and the Buffer<GenMessage> which is the inmemory struct representation, taking care of loopback of global and addressed messages, as well as filtering incoming addressed messages.

Fragmented messages are not yet supported. The datagram service is not yet supported. Routing between a CAN interface and a TCP interface is not yet supported.

Unit tests are added for the following schemes:
- static parsing and rendering functions; sequence generator
- the transmit flow communicating with an FD.
- the receive flow reading from an FD using select; with various alignment cases as it is reassembling a single string from the FD.
- the interface, for sending and receiving addressed and global messages, including tests for loopback and filtering noninteresting addressed messages
- the interface handling of VerifyNodeID messages.
- a TcpIf and a loopback socket connected as a client, sending and receiving wire payload directly on the fd.
- a TcpIf as a hub, with three loopback sockets connected as clients and each having its own TcpIf instance, sending GenMessages between each other using the high-level APIs. 

----

* Revert "Revert "TCP interface""

This reverts commit eb835d38815edc734cc11a0d318fd45192b6f309.

# Conflicts:
#	src/openlcb/sources

* Fix build breakages.