Improves performance of RefreshLoop object (#270)
by shortcutting the yield to the Executor in case the pollng
target has called the Notifiable inline. In this case there
is nothing to wait for, so we can iterate through the polling
members faster.

This avoids a problem where the CPU cost of executor yields can
exceed the amount of time between the polling loops, causing 100%
CPU usage just from the RefreshLoop. This problem could arise when
there are many polling members.
---
I repeatedly encountered a scenario during the loadtesting of the ESP32, when OpenMRN thread consumed 100% of the CPU, and the only actual flow that was running was the RefreshLoop. I believe the root cause of this problem was that there were many polling objects, and after K (=5) calls the Executor goes into and out of select(). Since the esp32 has a very expensive select implementation, we were paying a lot of costs just for the polling loop.
This is one of the pending fixes for that problem.
Another pending fix is that I'm adjusting the value of K for the esp32.