esp32: pin openmrn task to core 0 (#273)
Esp32: pins executor thread to core 0. This avoid starving the loop() task
when the executor thread is busy but the kernel scheduler "forgets" it on core 1.

The esp32 FreeRTOS multi-processor scheduler is not very smart in the presence of pinned tasks. It is possible to get a priority inversion when a non-pinned higher-priority task gets scheduled onto the core where there is a pinned lower-priority task. In this case the other core could go down to as low as running the idle task instead of the higher-priority task being moved to the idle core and the lower priority task running on its pinned core.

---

* Esp32: pins executor thread to core 0. This avoid starving the loop() task
when the executor thread is busy but the kernel scheduler "forgets" it on core 1.

* fix style

* Fix comment.