Limits buffer memory between GC bridge and write flow. (#271)
Adds LimitedPool, an implementation of a BufferPool that allows only a fixed number of buffers to be allocated.

After this is exhausted, further allocations get blocked until some earlier buffer gets freed. This is similar to FixedPool except the memory is allocated and returned to mainBufferPool, so the overhead on a per-instance basis is much lower.

Adds LimitedPool to the BufferPort component which is responsible for sending gridconnect packets to the Hub's write flow. A link-time configuration option defines whether we should limit the number of filled-in packets sent to the write flow.

Without this limitation if the write flow gets backed up, there is an unlimited amount of RAM we use up in the queue of Buffers going out of the BufferPort and waiting on the input of the WriteFlow. With this limit the batching size will automatically adjust to the throughput that the WriteFlow can sustain.

---

* Adds LimitedPool, an implementation of a BufferPool that allows only a ficed number of buffers to be allocated.
After this is exhausted, further allocations get blocked until some earlier buffer gets freed.
This is similar to FixedPool except the memory is allocated and returned to mainBufferPool, so
no the overhead on a per-instance basis is much lower.

Adds LimitedPool to the BufferPort component which is responsible for sending gridconnect packets to the
Hub's write flow. A link-time configuration option defines whether we should limit
the number of filled-in packets sent to the write flow.

* FIx comments and style.

* Fix style.

* Adds comments and removes redundant code.

* FIx typo in comment.

* fix comment typo.