increasing stack for esp32 socket_listener to 2kb
As part of switching to socket_listener for the JMRI connection in
DCC++ESP32 I found it crashing with stack canary violations. Increasing
the stack from 1.5kb to 2kb resolved this issue. There is a 512b stack
size bump in all built-in tasks when CONFIG_NEWLIB_NANO_FORMAT is not
enabled (disabled by default), this is in addition to the stack used by
the task.