Esp select cleanup (#285)
Removing an unused var and fixing a log statement for the select code for the esp32 as well as bumping default stack size to 2kb for socket_listener (there is a 512b "pad" required by esp32 housekeeping)

---

* Removing unused static var and fixing LOG data type to resolve compiler
warnings.

* increasing stack for esp32 socket_listener to 2kb

As part of switching to socket_listener for the JMRI connection in
DCC++ESP32 I found it crashing with stack canary violations. Increasing
the stack from 1.5kb to 2kb resolved this issue. There is a 512b stack
size bump in all built-in tasks when CONFIG_NEWLIB_NANO_FORMAT is not
enabled (disabled by default), this is in addition to the stack used by
the task.