handle zero byte internal config file (#307)
Fixes  #306, when the file exists but is zero bytes we can't read from it to check the version so it should treated as a reset instead.

===

* handle zero byte internal config file

fix for #306, when the file exists but is zero bytes we can't read from it to check the version so it should treated as a reset instead.

* review comments

Adjusted the short config file check to look for size shorter than version().end_offset(). Added VERBOSE level log messages to short file and version mismatch.