adding loop_executor() option (#309)
Adding the loop_executor() method which can be used to donate the current running thread to the executor. WDT check is in place for ESP32 and when WDT is active it will warn rather than cause fatal crash.

==

* adding loop_executor() option

Adding the loop_executor() method which can be used to donate the current running thread to the executor. WDT check is in place for ESP32 and when WDT is active it will warn rather than cause fatal crash.

* Update OpenMRNLite.h

missed update to comment for extra condition check

* review comments

call stack_->loop_executor() instead of stack_->executor()->loop...

* avoid second call to start_stack

Going back to original code in loop_executor() method and routing thread_entry to call loop_executor() for reduction of duplicated steps.