Adding Protection for the Executor add_from_isr() call.
On the ESP32 it is not guaranteed that the add_from_isr() call is protected by a critical section and thus it is necessary to invoke the insert() method instead of insert_locked().

The Atomic impl for the ESP32 has been updated to add checks for execution within an ISR context and calling the ISR variant if required.