Updating constants.hxx to fix issue on esp32
Using the datatype as "void" and _sym_XXX as a function causes the xtensa GCC compiler to generate code which is unable to successfully evaluate config_XXX() == CONSTANT_TRUE as boolean true even when config_XXX() returns CONSTANT_TRUE when printed. By switching the datatype to int and treating it as a variable instead, the compiler generates the expected equality test.

The previous PR for this used constant values of 4 and 8 which also worked but simply changing the declaration of the symbol is a smaller change.