Lower default priority for OpenMRN tasks on ESP32 (#339)
Lower the default priority of OpenMRN tasks on the ESP32 by one, this is to ensure that none of the OpenMRN tasks block the TCP/IP task from timely execution. Further refinements/documentation of task priorities will be done in a subsequent PR.

===

* Lower default priority for OpenMRN tasks on ESP32

Lower the default priority of OpenMRN tasks on the ESP32 by one, this is to ensure that none of the OpenMRN tasks block the TCP/IP task from timely execution. Further refinements/documentation of task priorities will be done in a subsequent PR.

* Update OpenMRNLite.h

correct the comment about the priority level.

* Update Esp32HardwareCanAdapter.hxx

adjust the RX_TASK_PRIORITY and TX_TASK_PRIORITY to be inline with the new value for OPENMRN_TASK_PRIORITY.