ESP-IDF v4 VFS support
IDF v4 has reworked the VFS layer in a breaking manner to IDF v3.x, specifically in the start_select and end_select VFS methods. The semaphore that IDF VFS passes to the esp_start_select API is now a structure containing extra metadata about the semaphore (LwIP or FreeRTOS owned). With this new data structure it is not necessary to obtain and trigger the LwIP semaphore directly.

Also added documentation for methods which were missing it and the cxx license comment copied from hxx file.