Initial pass at updates for the ESP32-S2.
The ESP32-S2 is a single core version of the ESP32 designed as a replacement for the ESP8266.

Key differences between ESP32 and ESP32-S2:
1. BLE, Ethernet, Hall Sensor support has been removed.
2. SRAM reduced from 520kb to 320kb.
3. ROM reduced from 448kb to 128kb.
4. Single CPU core.
5. 43 usable IO pins vs 34 IO pins.

The CAN driver is currently not supported on the ESP32-S2 but a variant of it will be available in ESP-IDF v4.2 (required for ESP32-S2 use anyway).