Refactored Esp32WiFiManager to support ESP-IDF v4.1+ while retaining most functionality for IDF v3.x.
IDF v4.0 introduced a new esp_event library to replace the event_loop system that was previously used for propagating WiFi events and event_loop was deprecated but still functional. In IDF v4.1 the event_loop code was made optional and is not included by default (arduino-esp32 is not including it in version 2.0 which will be coming soon).

Removed non-managed WiFi since it doesn't make as much sense as originally.

Removed the event callback support since it was never released in OpenMRNLite and will not compile in arduino-esp32 v2.0. The replacement is to hook into the esp_event system directly for the required events.