Reworked Esp32Gpio to allow usage of MultiConfiguredPC.
MultiConfiguredPC allows dynamic adjustment of pins to be input or output. If a pin that is input only is attempted to be reconfigured as output it will result in an error being asserted.

On the ESP32 pins from 0..33 are suitable for outputs. Note that pins 6-11 and 24 are not usable.

On the ESP32-S2 pins from 0..45 are suitable for outputs. Note that pins 22-24 and 26-32 are not usable.