Adds a single-file implementation of the newlib file APIs for Arduino. (#408)
The implementation forwards the calls to Arduino-STM32's
eeprom library. This is the bare minimum to support OpenMRN's
use of config file as the storage mechanism.

===

* Adds feature macro OPENMRN_HAVE_POSIX_FD.
Replaces uses of ifdef ARDUINO with feature macros.

* Adds a single-file implementation of the newlib file APIs.
The implementation forwards the calls to Arduino-STM32's
eeprom library. This is the bare minimum to support OpenMRN's
use of config file as the storage mechanism.

* Adds comment.

* Changes the constant type to off_t.