Adds a partial implementation of VirtualMemorySpace (#418)
The VirtualMemorySpace is an implementation of a MemorySpace interface which does not have a linear storage space behind. Instead, each variable is registered with a callback to write and read. This allows various advanced features, such as remapping address space, creating holes, really huge repetitions, etc.
This is the first cut of an implementation that is not feature complete, only one data type is supported, and has todos left.

===

* Adds VirtualMemorySpace
a memory space where the values are not stored in a
contiguous storage area but are read and written via callbacks.

* Adds comments.

* Adds more tests.

* Adds another test.

* Fix copyright date.

* fix typos.

* Adds unit tests for asynchronous execution on the memory space. Fixes a respective bug in the implementation.

* fully parenthesizes complex expressions.