Moving socket listener stack size and backlog count to constants. (#423)
As we discussed over email, the default number of sockets available for use is 10 with a max of 16. With the default backlog size of five it leaves only five additional sockets before errors will be reported by LwIP. ESP-IDF examples limit the backlog to one and thus that is the default with this PR.

This PR also includes optional naming of the SocketListener thread (useful when more than one listener is in use)

Fixes #256

===

* Moving socket listener stack size and backlog count to constants.

Fixes https://github.com/bakerstu/openmrn/issues/256

* Update socket_listener.cxx

Moved include references per feedback.