Refactors all usages of strncpy.
Adds a new function str_populate that invokes strncpy in length-safe manner,
ensuring that a terminating null is present even if the string is truncated.
Using C++ templates this also allows the caller not to need to specify the
length of the char[] array, which is safer.

Replaces all calls to strncpy that were writing into a fixed size C array
with the new function.