Adds support for using the stored setting for AP/STA mode. (#436)
Adds WlanRole::DEFAULT as an argument to the start() method that allows
to use the AP/STA setting from the NWP stored configuration instead of
from the application.

Adds wlan_set_role to override such configuration.

Fixes a bug where the IP address handed out via DHCP was incorrectly
saved in the AP's own IP address variable. This caused the "is local
connection" check in SocketClient to misbehave in AP mode.