Stores reserved aliases in the local alias cache (#467)
Updates the code and API of the AliasAllocator to store the list of aliases that are reserved but unused in the local_aliases() cache. This allows a large number of aliases to be pre-allocated using the BulkAliasAllocator without incurring extra memory overhead for keeping a list of Buffer<> objects in a queue. 
This PR makes the output of the BulkAliasAllocator be usable by the IfCan send flows.
The API contract of for alias allocation between the IfCan send flow and the AliasAllocator changes. Previously it was using a QAsync; how there is a direct sync/async API exposed on the AliasAllocator object. This new API simplifies the interaction because the checking for conflicts on reserved aliases is not the responsibility of the IfCan flow anymore but the AliasAllocator.

There is a feature that gets impacted by this change: to keep a certain number of reserved aliases around at all times even if they are not in use by any local node ID. The storage and API to trigger this feature was in the queued object that does not exist anymore. The feature is now triggered instead by an OVERRIDE_CONST. The default behavior will also change from true to false: previously an OpenMRN IO board would allocate two aliases and use only one of them; after this PR only one alias will be allocated and used by a simple IO board. Applications that use many virtual nodes should set the OVERRIDE_CONST instead.

Adds an API to the AliasCache to walk entries in node-ID-ascending order.

==

* start adding code for storing the reserved aliases differently.

* Adds a feature to call upper bound on the alias cache.

* Rewrites the accessor API for how to get allocated aliases from AliasAllocator.

* Updates the API how the reserved aliases are accessed.

* Makes create_allocated_alias test helper not require expect_next_alias_allocation call right afterwards.

* Removes unneeded alias injection.

* Fixes crashing bugs.

* removes unneeded call to create alias.

* give more sleep time to reduce flakiness.

* Adjust implementation to the new alias allocator behavior.

* TEST add log

* Adds a parameter to control whether a reserved but
unused alias will be allocated by the stack or not.

* Implements feature to keep a reserved but unused alias in the cache.

* Adds more comments to this test case.

* switch to using the alloc_result call on the passed-in executor.

* Adjusts the alias allocator tests for the new APIs.

* Fixes whitespace.

* Fix temporarily changed test APIs.

* Adds helpers to run code inline with class construction order.

* DEBUG statements.

* Adds a useful debugging facility.

* DEBUG Unnecessary debugging facilities.

* Adds a shorter alternative to run_x.

* Revert "DEBUG statements."

This reverts commit 00b0edd65c029ed0d7cc9bdaf3b55dbc8c81f0c4.

* Revert "DEBUG Unnecessary debugging facilities."

This reverts commit 2a7bbc20221172b0c546e14fbcbadeb8b918697b.

* Fixes comments.

* fix whitespace.