Adjust the ESP-IDF version detection logic.
The original logic worked with ESP-IDF v3.2 and v4.0+ due to `ESP_IDF_VERSION_MAJOR` only being defined in v4.0+. With arduino-esp32 v1.0.5 picking up ESP-IDF v3.3 this is no longer working due to ESP-IDF v3.3 now defining this constant.

Shifting to use __has_include to switch between the new and old include paths with fallback to use old paths when __has_include is not usable (though it always should be with GCC). This will need to be revisited as part of the ESP-IDF v4.x rework that I have in progress since there are a few other breaking API changes starting with ESP-IDF v4.1.