Fix incorrect global AME replies. (#480)
Fixes a bug where a global AME request was also getting replies for reserved
but unallocated aliases. These carry an invalid node ID from the local alias
cache and should not be sent to the bus.

Adds reserved aliases to the test cases and to test that no additional packets
show up.